+++ agent-bootstrap/src/main/java/io/opentelemetry/auto/config/Config.java
@@ -71,8 +71,9 @@ public class Config {
   public static final String RUNTIME_CONTEXT_FIELD_INJECTION =
       "trace.runtime.context.field.injection";
 
-  public static final String LOGS_INJECTION_ENABLED = "logs.injection.enabled";
-  public static final String LOGS_EVENTS_THRESHOLD = "logs.events.threshold";
+  public static final String LOG_INJECTION_ENABLED = "log.injection.enabled";
+  public static final String EXPERIMENTAL_LOG_CAPTURE_THRESHOLD =
+      "experimental.log.capture.threshold";
 
   private static final boolean DEFAULT_TRACE_ENABLED = true;
   public static final boolean DEFAULT_INTEGRATIONS_ENABLED = true;
@@ -89,7 +90,8 @@ public class Config {
   private static final boolean DEFAULT_DB_CLIENT_HOST_SPLIT_BY_INSTANCE = false;
   private static final int DEFAULT_SCOPE_DEPTH_LIMIT = 100;
 
-  public static final boolean DEFAULT_LOGS_INJECTION_ENABLED = false;
+  public static final boolean DEFAULT_LOG_INJECTION_ENABLED = false;
+  public static final String DEFAULT_EXPERIMENTAL_LOG_CAPTURE_THRESHOLD = null;
 
   private static final String SPLIT_BY_SPACE_OR_COMMA_REGEX = "[,\\s]+";
 
@@ -112,7 +114,7 @@ public class Config {
   @Getter private final Integer scopeDepthLimit;
   @Getter private final boolean runtimeContextFieldInjection;
 
-  @Getter private final boolean logsInjectionEnabled;
+  @Getter private final boolean logInjectionEnabled;
 
   // mapping of threshold values to different logging frameworks:
   //
@@ -128,7 +130,7 @@ public class Config {
   // | FINER        | FINER   | DEBUG   | DEBUG  |
   // | TRACE/FINEST | FINEST  | TRACE   | TRACE  |
   // | ALL          | ALL     | ALL     | ALL    |
-  @Getter private final String logsEventsThreshold;
+  @Getter private final String experimentalLogCaptureThreshold;
 
   @Getter private final String traceAnnotations;
 
@@ -184,10 +186,12 @@ public class Config {
         getBooleanSettingFromEnvironment(
             RUNTIME_CONTEXT_FIELD_INJECTION, DEFAULT_RUNTIME_CONTEXT_FIELD_INJECTION);
 
-    logsInjectionEnabled =
-        getBooleanSettingFromEnvironment(LOGS_INJECTION_ENABLED, DEFAULT_LOGS_INJECTION_ENABLED);
+    logInjectionEnabled =
+        getBooleanSettingFromEnvironment(LOG_INJECTION_ENABLED, DEFAULT_LOG_INJECTION_ENABLED);
 
-    logsEventsThreshold = getSettingFromEnvironment(LOGS_EVENTS_THRESHOLD, null);
+    experimentalLogCaptureThreshold =
+        getSettingFromEnvironment(
+            EXPERIMENTAL_LOG_CAPTURE_THRESHOLD, DEFAULT_EXPERIMENTAL_LOG_CAPTURE_THRESHOLD);
 
     traceAnnotations = getSettingFromEnvironment(TRACE_ANNOTATIONS, DEFAULT_TRACE_ANNOTATIONS);
 
@@ -244,10 +248,12 @@ public class Config {
         getPropertyBooleanValue(
             properties, RUNTIME_CONTEXT_FIELD_INJECTION, parent.runtimeContextFieldInjection);
 
-    logsInjectionEnabled =
-        getPropertyBooleanValue(properties, LOGS_INJECTION_ENABLED, parent.logsInjectionEnabled);
+    logInjectionEnabled =
+        getPropertyBooleanValue(properties, LOG_INJECTION_ENABLED, parent.logInjectionEnabled);
 
-    logsEventsThreshold = properties.getProperty(LOGS_EVENTS_THRESHOLD, parent.logsEventsThreshold);
+    experimentalLogCaptureThreshold =
+        properties.getProperty(
+            EXPERIMENTAL_LOG_CAPTURE_THRESHOLD, parent.experimentalLogCaptureThreshold);
 
     traceAnnotations = properties.getProperty(TRACE_ANNOTATIONS, parent.traceAnnotations);
 
@@ -635,7 +641,7 @@ public class Config {
       if (fileReader != null) {
         try {
           fileReader.close();
-        } catch (IOException ioe) {
+        } catch (final IOException ioe) {
           log.error("Configuration file '{}' was not closed correctly.", configurationFilePath);
         }
       }
