+++ instrumentation/instrumentation.gradle
@@ -90,7 +90,7 @@ configurations {
 }
 
 // need to perform shading in two steps in order to avoid shading java.util.logging.Logger
-// in the java-util-logging-events instrumentation since that instrumentation needs to
+// in the java-util-logging instrumentation since that instrumentation needs to
 // reference unshaded java.util.logging.Logger
 // (java.util.logging.Logger shading is not needed in any of the instrumentation modules,
 // but it is needed for the dependencies, e.g. guava, which use java.util.logging.Logger)
@@ -102,7 +102,7 @@ task shadowJarStep1(type: com.github.jengelman.gradle.plugins.shadow.tasks.Shado
   configurations = [project.configurations.runtime]
 
   dependencies {
-    exclude(project(':instrumentation:java-util-logging-events'))
+    exclude(project(':instrumentation:java-util-logging'))
   }
 
   // rewrite dependencies calling Logger.getLogger
