+++ instrumentation/log4j-2.0/src/test/groovy/Log4jThreadContextTest.groovy
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2020, OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import io.opentelemetry.auto.test.log.injection.LogContextInjectionTestBase
+import org.apache.logging.log4j.ThreadContext
+import org.junit.Ignore
+
+// FIXME this instrumentation relied on scope listener
+@Ignore
+class Log4jThreadContextTest extends LogContextInjectionTestBase {
+
+  @Override
+  def put(String key, Object value) {
+    return ThreadContext.put(key, value as String)
+  }
+
+  @Override
+  def get(String key) {
+    return ThreadContext.get(key)
+  }
+}
