+++ instrumentation/log4j-2.0/src/test/resources/log4j2.xml
