--- instrumentation/log4j1/log4j1.gradle
@@ -1,27 +1,0 @@
-apply from: "${rootDir}/gradle/java.gradle"
-
-ext {
-  log4jVersion = '1.2.17'
-}
-
-muzzle {
-  pass {
-    group = 'log4j'
-    module = 'log4j'
-    versions = '(,)'
-  }
-}
-
-configurations {
-  // In order to test the real log4j library we need to remove the log4j transitive
-  // dependency 'log4j-over-slf4j' brought in by :testing which would shadow
-  // the log4j module under test using a proxy to slf4j instead.
-  testCompile.exclude group: 'org.slf4j', module: 'log4j-over-slf4j'
-
-  // See: https://stackoverflow.com/a/9047963/2749853
-  testCompile.exclude group: 'javax.jms', module: 'jms'
-}
-
-dependencies {
-  testCompile group: 'log4j', name: 'log4j', version: log4jVersion
-}
