+++ settings.gradle
@@ -82,7 +82,7 @@ include ':instrumentation:java-concurrent:kotlin-testing'
 include ':instrumentation:java-concurrent:scala-testing'
 include ':instrumentation:java-concurrent:akka-testing'
 include ':instrumentation:java-concurrent:akka-2.5-testing'
-include ':instrumentation:java-util-logging-events'
+include ':instrumentation:java-util-logging'
 include ':instrumentation:jaxrs-1.0'
 include ':instrumentation:jaxrs-2.0'
 include ':instrumentation:jaxrs-2.0:jaxrs-2.0-jersey-2.0'
@@ -101,13 +101,9 @@ include ':instrumentation:jsp-2.3'
 include ':instrumentation:kafka-clients-0.11'
 include ':instrumentation:kafka-streams-0.11'
 include ':instrumentation:lettuce-5.0'
-include ':instrumentation:log4j-events-1.1'
-// FIXME this instrumentation relied on scope listener
-// include ':instrumentation:log4j1'
-include ':instrumentation:log4j-events-2.0'
-// FIXME this instrumentation relied on scope listener
-// include ':instrumentation:log4j2'
-include ':instrumentation:logback-events-1.0'
+include ':instrumentation:log4j-1.1'
+include ':instrumentation:log4j-2.0'
+include ':instrumentation:logback-1.0'
 include ':instrumentation:mongo-3.1'
 include ':instrumentation:mongo-3.7'
 include ':instrumentation:mongo-async-3.3'
